Backend : Add ReviewImage

+++ src/main/java/com/example/demo/repository/ReviewRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review,Integer>{
-    List<Review> findByPlacePlaceID(Long placeId);    
+    List<Review> findByPlacePlaceID(int placeId);    
 }
